requirements.txt
annotated-types==0.7.0
anyio==4.10.0
boto3==1.40.32
botocore==1.40.32
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.2
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
mangum==0.19.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moto==5.1.12
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
responses==0.25.8
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
s3transfer==0.14.0
sentry-sdk==2.38.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
xmltodict==1.0.0
